chore: Update language options in index.html
The commit updates the language options in the index.html file to include Albanian and English languages. This change allows users to select from a wider range of language options when using the Text To Speech functionality.

Note: This commit message follows the conventional format of starting with a type (chore) followed by a brief description of the changes.
Signed-off-by: Pito The Dev <shpetim.h@gmail.com>