chore: Fix file cleanup in app.py
This commit fixes a bug in the `app.py` file where the temporary audio file was not being properly closed and removed after it was sent as a response. The bug is fixed by adding a call to `os.remove()` to delete the temporary file before the response is returned.

Note: This commit message follows the conventional format of starting with a type (chore) followed by a brief description of the changes.
Signed-off-by: Pito The Dev <shpetim.h@gmail.com>