chore: Update title and labels in index.html
This commit updates the title and labels in the index.html file to use Albanian language. The title is changed to "Konverto Tekstin në Zë" and the labels for text, voice, and language are also translated to Albanian. This change improves the user experience for Albanian-speaking users.

Note: This commit message follows the conventional format of starting with a type (chore) followed by a brief description of the changes.
Signed-off-by: Pito The Dev <shpetim.h@gmail.com>